feat: Enhance authentication and authorization mechanisms across the application

+++ src/app/api/coupons/edit/[_id]/route.ts
@@ -1,8 +1,15 @@
 
-import { routeHandlerWrapper } from "@/action";
+import { checkAdminUserRequest, routeHandlerWrapper } from "@/action";
 import { Coupon } from "@/models/coupon.models";
+import { NextResponse } from "next/server";
 
 export const PATCH = routeHandlerWrapper(async (req: Request, { params }: { params: Promise<{ _id: string }> }) => {
+    const user = await checkAdminUserRequest()
+
+    if (!user) {
+        return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+
     const { _id } = await params;
     const body = await req.json();
 
@@ -21,6 +28,12 @@ export const PATCH = routeHandlerWrapper(async (req: Request, { params }: { para
 });
 
 export const DELETE = routeHandlerWrapper(async (req: Request, { params }: { params: Promise<{ _id: string }> }) => {
+    const user = await checkAdminUserRequest()
+
+    if (!user) {
+        return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+
     const { _id } = await params;
 
     const coupon = await Coupon.findByIdAndDelete(_id);
