+++ src/app/api/courses/admin/edit/[_id]/route.ts
@@ -1,13 +1,26 @@
-import { routeHandlerWrapper } from "@/action";
+import { checkAdminUserRequest, routeHandlerWrapper } from "@/action";
 import { Course } from "@/models/course.models";
+import { NextResponse } from "next/server";
 
 export const GET = routeHandlerWrapper(async (request: Request, { params }: { params: Promise<{ _id: string }> }) => {
+    const user = await checkAdminUserRequest()
+
+    if (!user) {
+        return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+
     const { _id } = await params;
     const course = await Course.findById(_id);
     return new Response(JSON.stringify(course), { status: 200 });
 });
 
 export const PATCH = routeHandlerWrapper(async (request: Request, { params }: { params: Promise<{ _id: string }> }) => {
+    const user = await checkAdminUserRequest()
+
+    if (!user) {
+        return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+
     const { _id } = await params;
     const courseData = await request.json();
     const updatedCourse = await Course.findByIdAndUpdate(_id, courseData, { new: true });
@@ -15,6 +28,12 @@ export const PATCH = routeHandlerWrapper(async (request: Request, { params }: { 
 });
 
 export const DELETE = routeHandlerWrapper(async (request: Request, { params }: { params: Promise<{ _id: string }> }) => {
+    const user = await checkAdminUserRequest()
+
+    if (!user) {
+        return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+    }
+
     const { _id } = await params;
     await Course.findByIdAndDelete(_id);
     return new Response(JSON.stringify({ message: "Course deleted successfully" }), { status: 200 });
